+++ Trabajo Practico 1/FuncionesAritmeticas.h
@@ -1,4 +1,4 @@
 int Suma(int numero1,int numero2);
 int Resta(int numero1, int numero2);
 int Multiplicacion(int factor1,int factor2);
-float Division(int dividendo,int divisor);
+float Division(float dividendo,float divisor);
